chore: refactor kube nodes column
Refactors the Kubernetes Nodes column using /columns and a props.ts.

Part of #11035.
